requirements.txt
pytest>=7.1.2
allure-pytest>=2.9.45
pydantic>=1.9.2
python-dotenv>=0.20.0
aiohttp>=3.8.1
pytest-asyncio>=0.19.0
bs4>=0.0.1
selenium>=4.0.0
lxml>=4.9.1
requests>=2.25.1